Refactor Pivot Table to allow widget for documents
